requirements.txt
accelerate
newsapi-python
pillow
psutil
pyaudio
pyautogui
pyttsx3
requests
sounddevice
speechrecognition

torch
transformers
vosk
wikipedia